Split out keyboard settings from Preferences
Introduce a new menu, "Keyboard settings", splitting it out of "Preferences".

Fixes #290.
