+++ src/renderer/components/MainMenu/KeyboardSettingsMenuItem.js
@@ -1,0 +1,43 @@
+// -*- mode: js-jsx -*-
+/* Chrysalis -- Kaleidoscope Command Center
+ * Copyright (C) 2018, 2019  Keyboardio, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from "react";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemIcon from "@material-ui/core/ListItemIcon";
+import ListItemText from "@material-ui/core/ListItemText";
+import USBIcon from "@material-ui/icons/Usb";
+import i18n from "../../i18n";
+
+export default function KeyboardSettingsMenuItem({
+  selected,
+  onClick,
+  className
+}) {
+  return (
+    <ListItem
+      button
+      selected={selected}
+      onClick={onClick}
+      className={className}
+    >
+      <ListItemIcon>
+        <USBIcon />
+      </ListItemIcon>
+      <ListItemText primary={i18n.app.menu.keyboardSettings} />
+    </ListItem>
+  );
+}
